+++ src/index.js
@@ -1,17 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import './App.css';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Login from "pages/Login";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+function App() {
+
+  return (
+    <BrowserRouter basename="/">
+      <Routes>
+        <Route exact path="/" element={<Login />} /> {/* => 로그인페이지 */}
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default App;
