+++ src/pages/login/style.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { MainContainer } from "./style";
+
+const Main = () => {
+
+    return (
+        <MainContainer>
+
+
+        </MainContainer>
+    )
+}
+
+export default Main;
